Discovery: Drop search results from previous queries
This patch adds a new optional parameter to the searchChannels function
to be able to drop search results for old searches.

The SearchPage component increments the searchQueryIndex for every
search, so if there's a race condition and a search query done before
ends after a second search query has been launched, it will be dropped
because the searchQueryIndex returned will be different from the one
specified in the new search query.

https://phabricator.endlessm.com/T33202