discovery: Ignore previous search promises
By keeping track of the last promise per each kind, we are able to
ignore the previous ones by not resolving them.

This fixes issues with old promises provoking inconsistent or
duplicate results coming from a previous search.

This is done in a new `searchChannelsOnce()` method which wraps the
previous `searchChannels()` so it can be easily replaced.

The previous searchChannels is now renamed to setSearchResults and the
fetch is moved to the wrapper, so it happens only once.

Also, mark all promises as "old" when the component is deactivated by
resetting the state.

https://phabricator.endlessm.com/T33202